refactor: move create_env_file_if_not_exists to utils and reuse in tests
