+++ backend/__main__.py
@@ -3,16 +3,7 @@ from pathlib import Path
 
 from backend.models import PuzzleData, Wordlist  # noqa: F401
 from backend.pages import Pages, PrintParams
-from backend.utils import Logger, Config, get_config, AppConfig, PuzzleConfig
-
-
-def create_env_file_if_not_exists():
-    env_path = Path(".env")
-    env_dist_path = Path(".env.dist")
-    if not env_path.exists():
-        with open(env_dist_path, "r") as ed_fd:
-            with open(env_path, "w") as fd:
-                fd.write(ed_fd.read())
+from backend.utils import Logger, Config, get_config, AppConfig, PuzzleConfig, create_env_file_if_not_exists
 
 
 def fix_puzzle_config(config: Config):
