+++ backend/tests/pages/test_sub_contents.py
@@ -5,9 +5,10 @@ import pytest
 from PIL import Image
 from backend.models import BoardImageEnum, DirectionEnum
 from backend.pages import SubContentsCell
+from ..test_utils import TestUtils
 
 
-class TestSubContentsCell:
+class TestSubContentsCell(TestUtils):
     """Test class for SubContentsCell"""
 
     @pytest.fixture
