+++ backend/utils/__init__.py
@@ -1,9 +1,20 @@
+from pathlib import Path
+
 from .config import Config, AppConfig, PrintConfig, AIConfig, PuzzleConfig  # noqa: F401
 from .logging import Logger  # noqa: F401
 
 
 profanity_list = None
 config: Config | None = None
+
+
+def create_env_file_if_not_exists():
+    env_path = Path(".env")
+    env_dist_path = Path(".env.dist")
+    if not env_path.exists():
+        with open(env_dist_path, "r") as ed_fd:
+            with open(env_path, "w") as fd:
+                fd.write(ed_fd.read())
 
 
 def get_profanity_list():
